requirements.txt
Flask
gunicorn
pandas
numpy
pillow
tensorflow
Pillow
Werkzeug